Fixed everything + culled unused dependencies

+++ wannapop/api/orders.py
@@ -1,5 +1,5 @@
 from . import api_bp
-from .errors import not_found, bad_request
+from .errors import not_found
 from ..models import Order, ConfirmedOrder
 from .helper_auth import token_auth
 from .helper_json import json_response, json_request, json_error_response
@@ -11,26 +11,28 @@ from flask import current_app
 @token_auth.login_required
 def api_order_add():
     user_id = token_auth.current_user().id
-    data = json_request(['product_id', 'buyer_id', 'offer', 'created'], False)
+    data = json_request(['product_id', 'offer'], False)
     order = Order.create(**data, buyer_id=user_id)
     return json_response(order.to_dict())
 
 @api_bp.route('/orders/<int:order_id>', methods = ['PUT'])
 @token_auth.login_required
 def api_order_edit(order_id):
+    user_id = token_auth.current_user().id
     order = Order.get(order_id)
-    if order:
-        
-        try:
-            user_id = token_auth.current_user().id
-            data = json_request(['product_id', 'offer', 'created'], False)
-        except Exception as e:
-            current_app.logger.debug(e)
-            return bad_request(str(e))
-        else:
+    if order and order.buyer_id == user_id:
+        is_confirmed = ConfirmedOrder.get(order_id)
+        if not is_confirmed:
+            data = json_request(['product_id', 'offer'], False)
             order.update(**data, buyer_id=user_id)
             current_app.logger.debug("UPDATED order: {}".format(order.to_dict()))
             return json_response(order.to_dict())
+        else:
+            current_app.logger.debug("Cannot edit order {}, already confirmed".format(order_id))
+            return json_error_response("403","Cannot edit this order, already confirmed")
+    elif order and order.buyer_id != user_id:
+        current_app.logger.debug("Current user is not author of order {}".format(order_id))
+        return json_error_response("403","Current user is not author of order")
     else:
         current_app.logger.debug("Order {} not found".format(order_id))
         return not_found("Order not found")
@@ -38,11 +40,24 @@ def api_order_edit(order_id):
 @api_bp.route('/orders/<int:order_id>', methods = ['DELETE'])
 @token_auth.login_required
 def api_order_delete(order_id):
+    user_id = token_auth.current_user().id
     order = Order.get(order_id)
-    if order:
-        order.delete()
-        current_app.logger.debug("CANCELLED order {}".format(order_id))
-        return json_response(order.to_dict())
+    if order and order.buyer_id == user_id:
+        is_confirmed = ConfirmedOrder.get(order_id)
+        if not is_confirmed:
+            order.delete()
+            current_app.logger.debug("CANCELLED order {}".format(order_id))
+            return json_response(order.to_dict())
+        else:
+            current_app.logger.debug("Cannot edit order {}, already confirmed".format(order_id))
+            return json_error_response("403","Cannot edit this order, already confirmed")
+    elif order and order.buyer_id != user_id:
+        current_app.logger.debug("Current user is not author of order {}".format(order_id))
+        return json_error_response("403","Current user is not author of order")
+    else:
+        current_app.logger.debug("Order {} not found".format(order_id))
+        return not_found("Order not found")
+
     
 @api_bp.route('/orders/<int:order_id>/confirmed', methods=['POST'])
 @token_auth.login_required
@@ -51,7 +66,7 @@ def api_order_confirm(order_id):
     is_confirmed = ConfirmedOrder.get(order_id)
     user=token_auth.current_user()
     if order:
-        if user.id != order.seller_id :
+        if user.id != order.buyer_id :
             return json_error_response("403", "User not authorized to accept this order")
         else:
             if is_confirmed:
@@ -66,13 +81,12 @@ def api_order_confirm(order_id):
 
 @api_bp.route('/orders/<int:order_id>/confirmed', methods=['DELETE'])
 @token_auth.login_required
-@token_auth.login_required
 def api_order_cancel(order_id):
-    user_id = token_auth.current_user().id
+    user = token_auth.current_user()
     order = ConfirmedOrder.get(order_id)
-    user=token_auth.current_user()
     if order:
-        if user.id != order.seller_id :
+        buyer_ref = Order.get(order_id).buyer_id
+        if user.id != buyer_ref:
             return json_error_response("403", "User not authorized to remove this order")
         else:
             order.delete()
